establish parent-ordering in `but-graph` (already)
This is important as it does things like fist-parent traversals,
which should really be traversing the first parent as defined
in the commit-graph, and not as defined by the insertion order
of the graph data structure.

----

Keep `parent_order` and queued `at_commit` as compact `u32` values instead
of `usize`. Both values are bounded by commit parent count / segment commit
index, but they sit in very hot graph traversal storage:

- `Edge` is stored in every petgraph edge slot.
- `Instruction::ConnectNewSegment` is part of every traversal `QueueItem`.

Using `usize` widened these layouts on 64-bit targets:

- `Edge`: 104 -> 112 bytes
- petgraph edge slot: 120 -> 128 bytes
- `Instruction`: 16 -> 24 bytes
- `QueueItem`: 256 -> 264 bytes

That was enough to cause a large regression on a large repository even with
`--no-post`, because traversal still queues, sorts, and builds graph edges.
With compact `u32` storage, the hot layouts return to their previous sizes:

- `Edge`: 104 bytes
- petgraph edge slot: 120 bytes
- `Instruction`: 16 bytes
- `QueueItem`: 256 bytes

Also add an allocation-free preflight before the parent-order edge rebuild so
the allocating rebuild path is skipped when outgoing edges are already in the
required insertion order.

Type-size checks used:

```sh
RUSTC_BOOTSTRAP=1 cargo rustc -p but-graph --release --lib -- \
  -Z print-type-sizes -C metadata=layout_probe > /tmp/but-graph-type-sizes.txt 2>&1
```

Benchmark command:

```sh
hyperfine --warmup 3 --runs 10 \
  './target/release/but-debug -C /Users/byron/dev-gb/gitlab graph --no-debug-workspace --no-post' \
  '/Users/byron/dev/github.com/gitbutlerapp/gitbutler/target/release/but-debug -C /Users/byron/dev-gb/gitlab graph --no-debug-workspace --no-post'
```

Observed after this change:

```
  ❯ hyperfine './target/release/but-debug -C ~/dev-gb/gitlab graph --no-debug-workspace --no-post' '/Users/byron/dev/github.com/gitbutlerapp/gitbutler/target/release/but-debug -C /Users/byron/dev-gb/gitlab graph --no-debug-workspace --no-post'
Benchmark 1: ./target/release/but-debug -C ~/dev-gb/gitlab graph --no-debug-workspace --no-post
  Time (mean ± σ):     787.2 ms ±   7.7 ms    [User: 597.9 ms, System: 185.6 ms]
  Range (min … max):   775.1 ms … 801.2 ms    10 runs

Benchmark 2: /Users/byron/dev/github.com/gitbutlerapp/gitbutler/target/release/but-debug -C /Users/byron/dev-gb/gitlab graph --no-debug-workspace --no-post
  Time (mean ± σ):     760.3 ms ±   4.6 ms    [User: 586.3 ms, System: 170.5 ms]
  Range (min … max):   752.3 ms … 766.8 ms    10 runs

Summary
  /Users/byron/dev/github.com/gitbutlerapp/gitbutler/target/release/but-debug -C /Users/byron/dev-gb/gitlab graph --no-debug-workspace --no-post ran
    1.04 ± 0.01 times faster than ./target/release/but-debug -C ~/dev-gb/gitlab graph --no-debug-workspace --no-post
```
`

The remaining gap is about `1.04x`, down from the previously observed larger
regression of 1.24x, and the hot traversal data layouts now match master.

+++ crates/but-graph/src/api.rs
@@ -50,6 +50,10 @@ impl Graph {
     /// `dst_commit_id` can be provided if the connection is to a future commit that isn't yet available
     /// in the `segment`. If `None`, it will be looked up in the `segment` itself.
     ///
+    /// `parent_order` is the 0-based position of `dst_commit_id` among the source commit's
+    /// parents. For a merge commit with parents `[A, B]`, the edge to `A` must use `0`,
+    /// and the edge to `B` must use `1`, even if traversal discovers `B` before `A`.
+    ///
     /// Return the newly added segment.
     pub fn connect_new_segment(
         &mut self,
@@ -58,7 +62,7 @@ impl Graph {
         dst: Segment,
         dst_commit: impl Into<Option<CommitIndex>>,
         dst_commit_id: impl Into<Option<gix::ObjectId>>,
-        parent_order: usize,
+        parent_order: u32,
     ) -> SegmentIndex {
         let dst = self.inner.add_node(dst);
         self.inner[dst].id = dst;
